Fix vector iterator bugs and add iterator constructor
Fix iterator but relating to not being able to properly deduce the end of the iterator.
The fix for this was to break out into a new base iterator type named parent_iterator which all iterator bases derive from. Parent iterator defines the fields all iterators have and the standard equivalency operators.